requirements.txt
streamlit
python-docx
pypdf
lucide-react
framer-motion
